cleaned up folder structured and names
