+++ lib/lqr.py
