+++ lib/madgwickahrs.py
@@ -16,8 +16,10 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
+import warnings
 import numpy as np
 import numbers
+from numpy.linalg import norm
 
 
 class Quaternion:
@@ -145,3 +147,127 @@ class Quaternion:
 
     def __array__(self):
         return self._q
+
+class MadgwickAHRS:
+    samplePeriod = 1/256
+    quaternion = Quaternion(1, 0, 0, 0)
+    beta = 1
+    zeta = 0
+
+    def __init__(self, sampleperiod=None, quaternion=None, beta=None, zeta=None):
+        """
+        Initialize the class with the given parameters.
+        :param sampleperiod: The sample period
+        :param quaternion: Initial quaternion
+        :param beta: Algorithm gain beta
+        :param beta: Algorithm gain zeta
+        :return:
+        """
+        if sampleperiod is not None:
+            self.samplePeriod = sampleperiod
+        if quaternion is not None:
+            self.quaternion = quaternion
+        if beta is not None:
+            self.beta = beta
+        if zeta is not None:
+            self.zeta = zeta
+
+    def update(self, gyroscope, accelerometer, magnetometer):
+        """
+        Perform one update step with data from a AHRS sensor array
+        :param gyroscope: A three-element array containing the gyroscope data in radians per second.
+        :param accelerometer: A three-element array containing the accelerometer data. Can be any unit since a normalized value is used.
+        :param magnetometer: A three-element array containing the magnetometer data. Can be any unit since a normalized value is used.
+        :return:
+        """
+        q = self.quaternion
+
+        gyroscope = np.array(gyroscope, dtype=float).flatten()
+        accelerometer = np.array(accelerometer, dtype=float).flatten()
+        magnetometer = np.array(magnetometer, dtype=float).flatten()
+
+        # Normalise accelerometer measurement
+        if norm(accelerometer) is 0:
+            warnings.warn("accelerometer is zero")
+            return
+        accelerometer /= norm(accelerometer)
+
+        # Normalise magnetometer measurement
+        if norm(magnetometer) is 0:
+            warnings.warn("magnetometer is zero")
+            return
+        magnetometer /= norm(magnetometer)
+
+        h = q * (Quaternion(0, magnetometer[0], magnetometer[1], magnetometer[2]) * q.conj())
+        b = np.array([0, norm(h[1:3]), 0, h[3]])
+
+        # Gradient descent algorithm corrective step
+        f = np.array([
+            2*(q[1]*q[3] - q[0]*q[2]) - accelerometer[0],
+            2*(q[0]*q[1] + q[2]*q[3]) - accelerometer[1],
+            2*(0.5 - q[1]**2 - q[2]**2) - accelerometer[2],
+            2*b[1]*(0.5 - q[2]**2 - q[3]**2) + 2*b[3]*(q[1]*q[3] - q[0]*q[2]) - magnetometer[0],
+            2*b[1]*(q[1]*q[2] - q[0]*q[3]) + 2*b[3]*(q[0]*q[1] + q[2]*q[3]) - magnetometer[1],
+            2*b[1]*(q[0]*q[2] + q[1]*q[3]) + 2*b[3]*(0.5 - q[1]**2 - q[2]**2) - magnetometer[2]
+        ])
+        j = np.array([
+            [-2*q[2],                  2*q[3],                  -2*q[0],                  2*q[1]],
+            [2*q[1],                   2*q[0],                  2*q[3],                   2*q[2]],
+            [0,                        -4*q[1],                 -4*q[2],                  0],
+            [-2*b[3]*q[2],             2*b[3]*q[3],             -4*b[1]*q[2]-2*b[3]*q[0], -4*b[1]*q[3]+2*b[3]*q[1]],
+            [-2*b[1]*q[3]+2*b[3]*q[1], 2*b[1]*q[2]+2*b[3]*q[0], 2*b[1]*q[1]+2*b[3]*q[3],  -2*b[1]*q[0]+2*b[3]*q[2]],
+            [2*b[1]*q[2],              2*b[1]*q[3]-4*b[3]*q[1], 2*b[1]*q[0]-4*b[3]*q[2],  2*b[1]*q[1]]
+        ])
+        step = j.T.dot(f)
+        step /= norm(step)  # normalise step magnitude
+
+        # Gyroscope compensation drift
+        gyroscopeQuat = Quaternion(0, gyroscope[0], gyroscope[1], gyroscope[2])
+        stepQuat = Quaternion(step.T[0], step.T[1], step.T[2], step.T[3])
+
+        gyroscopeQuat = gyroscopeQuat + (q.conj() * stepQuat) * 2 * self.samplePeriod * self.zeta * -1
+
+        # Compute rate of change of quaternion
+        qdot = (q * gyroscopeQuat) * 0.5 - self.beta * step.T
+
+        # Integrate to yield quaternion
+        q += qdot * self.samplePeriod
+        self.quaternion = Quaternion(q / norm(q))  # normalise quaternion
+
+    def update_imu(self, gyroscope, accelerometer):
+        """
+        Perform one update step with data from a IMU sensor array
+        :param gyroscope: A three-element array containing the gyroscope data in radians per second.
+        :param accelerometer: A three-element array containing the accelerometer data. Can be any unit since a normalized value is used.
+        """
+        q = self.quaternion
+
+        gyroscope = np.array(gyroscope, dtype=float).flatten()
+        accelerometer = np.array(accelerometer, dtype=float).flatten()
+
+        # Normalise accelerometer measurement
+        if norm(accelerometer) is 0:
+            warnings.warn("accelerometer is zero")
+            return
+        accelerometer /= norm(accelerometer)
+
+        # Gradient descent algorithm corrective step
+        f = np.array([
+            2*(q[1]*q[3] - q[0]*q[2]) - accelerometer[0],
+            2*(q[0]*q[1] + q[2]*q[3]) - accelerometer[1],
+            2*(0.5 - q[1]**2 - q[2]**2) - accelerometer[2]
+        ])
+        j = np.array([
+            [-2*q[2], 2*q[3], -2*q[0], 2*q[1]],
+            [2*q[1], 2*q[0], 2*q[3], 2*q[2]],
+            [0, -4*q[1], -4*q[2], 0]
+        ])
+        step = j.T.dot(f)
+        step /= norm(step)  # normalise step magnitude
+
+        # Compute rate of change of quaternion
+        qdot = (q * Quaternion(0, gyroscope[0], gyroscope[1], gyroscope[2])) * 0.5 - self.beta * step.T
+
+        # Integrate to yield quaternion
+        q += qdot * self.samplePeriod
+        self.quaternion = Quaternion(q / norm(q))  # normalise quaternion
