--- madgwick_py/__init__.py
@@ -1,19 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-    Copyright (c) 2015 Jonas Böer, jonas.boeer@student.kit.edu
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU Lesser General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU Lesser General Public License for more details.
-
-    You should have received a copy of the GNU Lesser General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-"""
-
-__all__ = ["madgwickahrs", "quaternion"]
