+++ setup/setup_mic.sh
