+++ setup/setup_speaker.sh
