+++ tests/test_led_strip.py
