+++ tests/test_speaker.py
