+++ tests/test_speaker_with_led.py
